test(ast-ngram): enforce file-length limit, fix naming, harden roundtrip
- Extract T5 Display formatting tests (5 tests, 54 lines) into
  ngram_display_tests.rs; wire via #[cfg(test)] #[path] in ngram.rs.
  ngram_tests.rs is now 396 lines, within the 400-line limit.
- Rename section T9b → T10 and shift T10-T15 → T11-T16 to restore the
  clean consecutive numeric sequence broken when T9b was added.
- Replace silent if-let skip in vocab_resolve_and_lookup_are_inverses
  with unwrap_or_else(panic!) so a missing fundamental kind causes a
  loud test failure instead of a vacuous pass.

applies ADR-001
